+++ src/store/errorMessageStore.ts
@@ -5,6 +5,8 @@ interface ErrorMessageStore {
   setMenuError: (error: string) => void;
   categoryError: string;
   setCategoryError: (error: string) => void;
+  loginError: string;
+  setLoginError: (error: string) => void;
 }
 
 export const useErrorMessageStore = create<ErrorMessageStore>((set) => ({
@@ -12,4 +14,6 @@ export const useErrorMessageStore = create<ErrorMessageStore>((set) => ({
   setMenuError: (error) => set({ menuError: error }),
   categoryError: '',
   setCategoryError: (error) => set({ categoryError: error }),
+  loginError: '',
+  setLoginError: (error) => set({ loginError: error }),
 }));
